res silenced some 64bit warnings in the Opcode plugin.
git-svn-id: http://crystal.svn.sourceforge.net/svnroot/crystal/CS/trunk@20088 8cc4aa7f-3514-0410-904f-f2cc9021211c

+++ include/ivaria/collider.h
@@ -143,7 +143,7 @@ struct iCollideSystem : public iBase
    * Get number of collision pairs in array.
    * \return the number of collision pairs.
    */
-  virtual int GetCollisionPairCount () = 0;
+  virtual size_t GetCollisionPairCount () = 0;
 
   /**
    * Reset the array with collision pairs. It is very important to call
